# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: |-
    REST API системы управления банковскими картами.

        Функциональность:
        - регистрация и аутентификация пользователей
        - управление банковскими картами
        - переводы между картами
        - заявки на активацию и блокировку карт
        - разграничение доступа по ролям (ADMIN / USER)

        Аутентификация осуществляется с помощью JWT (Bearer Token).
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  /api/users/{id}:
    get:
      tags:
      - Users
      summary: Получить пользователя по ID (только ADMIN)
      operationId: getById
      parameters:
      - name: id
        in: path
        description: ID пользователя
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
      security:
      - bearerAuth: []
    put:
      tags:
      - Users
      summary: Обновить пользователя (ADMIN)
      description: "Позволяет изменить имя, фамилию и роль пользователя"
      operationId: updateUser
      parameters:
      - name: id
        in: path
        description: ID пользователя
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
        required: true
      responses:
        "404":
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав ADMIN
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Пользователь обновлён
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
      security:
      - bearerAuth: []
  /api/users/registration:
    post:
      tags:
      - Users
      summary: Регистрация нового пользователя
      description: Создаёт пользователя с ролью USER
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Пользователь успешно создан
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          description: Некорректные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
        "409":
          description: Пользователь с таким email уже существует
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
      security:
      - bearerAuth: []
...
  /api/users:
    get:
      tags:
      - Users
      summary: Получить список пользователей (только ADMIN)
      description: "Фильтрация по email, имени, роли и дате создания"
      operationId: getUsers
      parameters:
      - name: email
        in: query
        description: Email пользователя
        required: false
        schema:
          type: string
      - name: firstname
        in: query
        description: Имя
        required: false
        schema:
          type: string
      - name: lastname
        in: query
        description: Фамилия
        required: false
        schema:
          type: string
      - name: role
        in: query
        description: Роль пользователя
        required: false
        schema:
          type: string
          enum:
          - ADMIN
          - USER
      - name: createdFrom
        in: query
        description: Дата создания ОТ
        required: false
        schema:
          type: string
          format: date-time
      - name: createdTo
        in: query
        description: Дата создания ДО
        required: false
        schema:
          type: string
          format: date-time
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав ADMIN
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Список пользователей
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUserDto"
      security:
      - bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    UpdateUserDto:
      type: object
      description: Данные для обновления пользователя (ADMIN)
      properties:
        firstname:
          type: string
          description: Имя пользователя
          example: Ivan
          minLength: 1
        lastname:
          type: string
          description: Фамилия пользователя
          example: Petrov
          minLength: 1
        role:
          type: string
          description: Роль пользователя
          enum:
          - ADMIN
          - USER
          example: ADMIN
      required:
      - firstname
      - lastname
      - role
    UserDto:
      type: object
      description: Информация о пользователе
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 10
        email:
          type: string
          description: Email пользователя
          example: user@example.com
        firstname:
          type: string
          description: Имя пользователя
          example: Ivan
        lastname:
          type: string
          description: Фамилия пользователя
          example: Petrov
        role:
          type: string
          description: Роль пользователя
          example: ROLE_USER
    RegisterDto:
      type: object
      description: Данные для регистрации пользователя
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
          example: new.user@example.com
          minLength: 1
        firstname:
          type: string
          description: Имя пользователя
          example: Ivan
          maxLength: 100
          minLength: 0
        lastname:
          type: string
          description: Фамилия пользователя
          example: Petrov
          maxLength: 100
          minLength: 0
        password:
          type: string
          description: Пароль пользователя
          example: StrongPassword123
          maxLength: 100
          minLength: 6
      required:
      - email
      - firstname
      - lastname
      - password
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageUserDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserDto"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT